UMTS RLC: Free with wmem what was allocated with wmem
Commit 2db3db56bfd1ba38c5f42015622dbc8530ed05e9 changed to using
wmem_file_scope memory for the frags, to prevent leaks caused by
dangling frags in malformed packets. The freeing functions need
to also call the matching wmem_free or else on file close there
can be complaints, even crashes with FORTIFY_SOURCE about freeing
the middle of a wmem_allocated block and possibly a double free.

This GHashtable could be changed to a autoreset wmem map, which
will be done in a follow up commit. (This is small in order to
make backporting simpler.)

Also avoid UB warnings in the case of memcpy'ing a NULL array
with a zero length (which is undefined according to the C standard.)


(cherry picked from commit a2c8ff7cb6b020b60adbd5a2d3e0111b26edd09f)

Co-authored-by: John Thacker <johnthacker@gmail.com>