epan: Reduce memory consumption with hostname lookups off
The get_hostname and get_hostname6 functions return const char*
that are directly from the cache of names. This has the effect
of caching a string (simply the standard string representation of
the address) even when network host name resolution is entirely
turned off.

The returned string, despite being a const char*, has a lifetime
shorter than that of the entire program, and could actually be
used after freed if not copied. (Not as commonly used, though.)

Add functions that take a wmem_allocator and return a copy of
the string in that scope. This allows avoiding adding to persistent
memory when host name lookups are completely disabled. It's slightly
slower, in that the string copy is somewhat more expensive than the
hash map lookup.

Use the function in the IP and IPv6 dissectors. Make up for the
performance impact and then some by only getting the hostname
and adding the FT_STRING hostname fields if the fields are referenced.
In all cases, the conditional hostname lookup is after the bytes have
already been retrieved from the tvb (perhaps to add the FT_IPv4
version of the same field), so this doesn't affect when any exceptions
would be thrown.

Together with disabling the conversation ID numbers for IP and IPv6
(commit a9b2be2df16d3f447), this allows filtering long lived captures
at the IP level (with upper level dissectors disabled) without memory
use increasing, and speeds up IP dissection.

Fix #20131. Ping #17776.