Diameter: Dissect User-Name as IMSI on SWx interface
3GPP TS 29.273i section 8 indicates that the User-Name AVP is the
IMSI on SWx, in the same way as S6a, S7a, etc.

The exception is for Command Code 305 (Push-Profile), which can
be either a single IMSI (in the case of a User Profile Update),
or a user list of a wildcard or IMSI digit prefixes in the case
of an HSS Reset Indication.