gsm_sim: Slice off the status word before dissecting command
The interpretation of the response data depends on the particular
command. For that reason, SIMtrace concatenates the response with
the request APDU. Slice off the status word from the TVB before
handing it to the command dissector, so that it isn't interpreted
as data.

While the reponse data is normally expected to be included in that
case, it might fail to be there if the command failed. Don't try
to add it and create a malformed packet in that case; it's more
useful to dissect the status word. (There probably should be an
expert info for abnormal status.)

Ping #16236. Ping #17299