SMB2: Fix a null pointer dereference
When one of the dissectors in the smb2_dissector array is
called (ed. this should probably be a dissector table),
si, the smb2_info_t is never NULL. However, si->saved might
be NULL.

Check for si->saved in one case. Fix #20149.

Remove some redundant checks for si when we only need to test
for si->saved, which quiets about a dozen Coverity warnings
about unnecessary extra null checks.