DHCP: Strengthen Alcatel-Lucent Option 43 heuristics
There are other vendors besides Alcatel-Lucent that for Option 43
use a generic encapsulated suboption structure like in Option 125.
If the Alcatel-Lucent Option 43 heuristic dissector sees a suboption
number that is used by Alcatel-Lucent Option 43 with a fixed length
but the actual length is wrong, reject it. It might be that of a
different vendor.

That prevents reporting errors and aborting dissection. It is not
the same as actually implementing heuristic dissectors for other
unknown vendors.

Ping #13953.