FTDI FT: Fix endpoint to interface mapping
The endpoint to interface mapping is broken since commit baf0bfc263ea
("USB: Store endpoint with direction flag in usb_conv_info"). Fix the
issue by using both endpoint number and direction flag in the switch
statement. Correct the endpoint IN and OUT comments to adhere to USB
specification naming (IN is from device to host, OUT is from host to
device).