RADIUS: Split out the illegal Ascend attributes in RFC 6929 space
Ascend products used a large number of RADIUS attributes in an
illegal manner. The subset of attributes in the IETF RFC 6929
extended attribute space are still fairly commonly used, especially
242.

By default we dissect those as the RFC 6929 extended attributes.
We have a preference to disable the RFC 6929 handling.

Including by default a dictionary with the Ascend illegal attributes
in the extended space has no effect when the preference is at its
default value, but causes the Ascend attributes to be used when
the preference is enabled. This makes it easier to use these attributes
by only changing preferences instead uncommenting lines in the
dictionary file.

The AVPs in the lower address space that have been assigned via
IETF Review are much less commonly used these days, and a user would
often not want to use the Ascend illegal values for those even if
using the Ascend AVPs in the extended address space. (The Reserved
for Private Use and Implementation Specific space perhaps are a
third category, but not handled here.)

Fix #15551