file: Fix find with no frame selected, wrap, and no match
When searching with no frame selected, the previous frame
number is set to 0 and the original frame data is NULL. If
wrapping is allowed, then the current frame and frame data will
never match those values, so the search will never end in failure.
That means that if no frame matches the search criterion, the
find will never stop unless the user manually stops it.

Don't wrap when no start packet is selected. If no start packet is
selected and searching backwards, set the first packet to test to
be at the end. Note that previously if no start packet was selected
searching backwards wouldn't work at all, which was presumably
unexpected.

Ensure we don't wrap twice (this check shouldn't be necessary, but
doesn't hurt in case there's some error elsewhere.)