test: Fix Sharkd TLS secrets test on Big-Endian
The file used, tls12-dsb.pcapng, has two used CLIENT_RANDOM
secrets, and the order in which they're output is related to
iterating over a hash table and is unpredictable. In particular,
it's different on Big Endian platforms than Little Endian.
Test for the Base64 of either of the two possibilities.

Ping #19965