TCP: Defer getting the window size until later
If the TCP header isn't long enough to contain the window size
(due to truncation), don't do sequence analysis (because we
need the window size for that) but otherwise defer retrieving
the window size and adjusting it until we're ready to add it to
the tree and we're done adding the stream number, flags, etc.

This allows a lot more information to be put in the tree in the
edge case where the TCP header is truncated to 14 octets instead
of 16 octets.

Fix #20155. Fix #16396.