Diameter: Dissect User-Name as EAP NAI on STa, SWm, S6b, etc.
Per 3GPP TS 29.273, on STa, SWm, S6b, SWa, and SWd when Command
Code 268 (Diameter-EAP) is used, the User-Name AVP "shall be represented
in NAI form as specified in IETF RFC 4282 [15] and shall be formatted
as defined in clause 19 of 3GPP TS 23.003 [14]. This IE shall include the
leading digit used to differentiate between authentication schemes,"

Dissect it as such. Note that on the those application IDs for other
command codes, the User-Name AVP does *not* contain the leading digit.