epan: Fixup adjusting length of FT_PROTOCOL value
Instead of hacking the setter to adjust the length by passing a
NULL tvb value as a flag, add a separate function for that.

There are some legitimate uses of setting a protocol_value_t with
a NULL tvb and we do not want to skip setting the string in that
case.

Fixup 51de43cfd2467f3e9f63fe0970ac9fb4553b3dd7.