epan: Increase maximum name for DNS names
Per RFC 1035, the total length of a domain name is 255 octets or
less. 63 octets or less is the maximum size of a single label, or
most other types of names we use (ports to services, names for
EUI-48, etc.) Increase the size of the array for resolved DNS names
from 64 to 256.

Fix #20145