Qt: Add X and Y axis context menus to I/O Graph
Add a context menu for the X axis that toggles the Time of Day setting.
Add a context menu for the Y axis that toggles the logarithmic scale
setting.

As with the "Enable legend" item, we might want to remove the
corresponding checkboxes from the main dialog window in order to
decrease the minimum width and make the I/O graph usable on smaller
resolutions (especially with some translations - #20147.)