wmem: Add wmem_map_find function
Add a function to search for a value in a map with that satisfies
a given function. This is O(N) to search the map, so it shouldn't
be routinely called for all values in the map (that's a sign that
the data structures need rethinking), but it can be useful.