TCP: check first if a change is going to be made in the first place
In all these blocks complex expressions are evaluated, but a change to
the tcp analysis data is only made when the frame was not yet visited.
Rather than doing all the work for nothing, first check if the frame
was visited, and only if not evaluate further.

had_acc_ecn_setup_syn and had_acc_ecn_setup_syn_ack tcp analysis data
fields not even took in consideration to only evaluate the first time
the frame is visited. These have now been included.