Qt: Prevent overflow in I/O Graph simple moving average
The warmup_interval, mvag_to_remove, and mavg_to_add variables
are always multiples of the fixed interval, and when retrieving
the item value, we divde by the fixed interval. There's no point
in doing that; just keep track of everything in term of indices.

In particular, this removes the possibility of overflow, which
could occur in 4.4 and later because the interval scaling changed
by a factor of 1000 in order to support intervals down to 1 microsecond.
( b46c1e9fc7ecb5f74c839dfca198e4022b2d2075 )

Also add a failure check if a negative index is passed into
get_io_graph_item instead of running off the left end of an array.

Fix #20163