Qt: Remove I/O Graph enableLegend checkbox in favor of context menus
The QCheckBox controls can't be shrunk by default and increase the
minimum width of the I/O Graph dialog. Remove the new checkbox
to enable and disable the legend in favor of the context menu
actions, which decreases the minimum width by 100 or so pixels
when English is used.

Also add the enable/disable Legend action to the Legend
context menu.

Ping #20147 (due to the longer translations in other languages,
we might need to remove other checkboxes, such as the log scale
and time of day, in favor of the context menus in order to get
the minimum width below 1024 or 800 pixels there.)