HTTP2: Fix decompression of single padded DATA frames
Workaround process_reassembled_data not truncating one-and-only-one
"reassembly" tvbs properly by slicing the padding off the tvb
immediately after calculating it.

Fix #20167

This can and should still get properly fixed in the reassembly
functions, but trimming the padding off immediately also simplifies
handling. (We could, as a result, reduce the number of times we
pass around the data length and offset, though not done here.)