Remove "outfile" pragma from flex files
Using this pragma is unnecessary and undesirable for the following
reasons:

1) It is surprising and unusual that a pragma inside the file overrides
the command-line options, usually command-line options take precedence
over hard-coded values, but not for flex.

2) Not all flex files in the tree use this "outfile" option pragma so
its use is inconsistent.

3) More importantly this ignores the command-line --output flag and forces
the output to be created on the current working directory with the specified
name. The fact that it works correctly at all with CMake in completely
incidental to an implementation detail, because CMake changes the working
directory to the correct output build directory. This is not a universal
equirement and may cause bizarre and difficult to troubleshoot failures
in different circumstances.

Replace "-o" with "--outfile" in the cmake file for better readability.