Qt: I/O Graph click and zoom only with LeftButton
Clicks other than LeftButton shouldn't go to the selected packet
(that's our behavior in, e.g., the FlowGraph.)

QCustomPlot controls dragging, and only performs it with the LeftButton.
Pressing and releasing other buttons shouldn't affect whether we're
dragging. We control zooming, so possibly we could allow it with some
other buttons that aren't RightButton.