QUIC: Track a list of server addresses and ports
In order to support migrating a server address when the client
Connection ID is zero-length, keep a list of server address/port pairs
and add to it when a new server address/port combination is used.
Check the entire list when determining if a frame is from the server side
of a connection. (This is only necessary with zero-length Connection
IDs and Stateless Reset tokens.)

RFC 9000 Section 9 implies that the client can only migrate a server to
the address and port given in the preferred_address QUIC transport parameter
in the TLS handshake, though we don't enforce or warn about that.
Note there exist clients that try to migrate to addresses obtained via other
means, such as DNS, and servers that will agree to migrate the connection.

Fix #20165