dot11decrypt: Fix Fast BSS Transition decryption on big-endian systems
The mdid field used as input for the PTK derivation for Fast BSS
Transition is incorrectly converted to little-endian byte order.
Due to this the derived PTK is invalid on big-endian systems,
making the decryption fail. This error also results in a few failed
tests on big-endian systems.

The mdid field is a 2-octet value, not a 16-bit value so byte order
conversion should not be applied. Fix by removing the byte order
conversion.

Ping #19965


(cherry picked from commit 361c5c856227d0e80987252c5661da8978d85351)

Co-authored-by: Mikael Kanstrup <mikael.kanstrup@gmail.com>