5CoRAP: Advance offset correctly when too many items
Even when not displaying the sub frame because too many (recursive)
subdevices have been dissected, the offset still needs to advance
by the header length as well as the length of the data size.

In a case where the data size is zero, not doing this can lead to
an infinite loop.

Fix #20176