epan: Fix length of returned fragment in one frag "reassembly" case
As an optimization, if the fragment_add_seq() functions are called
in a way that indicates that there is a "reassembly" with only one
frag, the data from the tvb is not copied to persistent memory.
Instead, when process_reassembled_data is called, the appropriate
subset of the passed tvbuff is returned back. (It would save even
more memory not to keep a record of the fragment, but perhaps
sometimes dissection logic is simpler this way.)

When storing information in the fragment head in that case, also store
the total length. Then, when returning the tvbuff subset, create a
subset of the proper length. The calling dissector probably has some way
to access that length, but this is surer.

See #20167