nbap: Fix a NULL pointer dereference in logging
This shows up with debugging enabled and probably on a second
pass. The conversation info isn't retrieved (or set up) when it
already exists, so we can't dereference it to log. (The compiler
warning about the variable being used uninitialized was relevant.)

This bug was dormant under not usually compiled debug code for
over a decade.

Fixup b205222b613d03f7f98ef3671c245cf41f1df135
Fix #20177