protobuf: Close repeated field JSON arrays followed by unknown fields
Repeated fields in protobuf, when expressed in JSON, use JSON arrays.
The JSON array cannot be closed until we either get a different field
type, or end the entire parsing (possibly by aborting.) There's one
check for ending the pending array if the last validly parsed field
was a repeated field.

However, when a repeated field is followed by an unknown field,
dissect_one_protobuf_field returns true, which results in
prev_field_desc being set to the null value, and the array not
being closed. Close a pending array if the new field is unknown.

Fix #20178.