elasticsearch: versions are really unsigned, fix format truncation
Elasticsearch is written in Java, so its ints (stored as varints)
are all signed. The version number is really unsigned (and the
signed bit should never be set anyway.) It's already added to the
tree cast as unsigned. Store it that way as well, so that when
performing div and mod to extract the components the result is
always non-negative (C and its silly truncated division).

This has the benefit of preventing format truncation warnings.

Ping #20175