dfilter: resolve unparsed for each node in a set
In check_relation_in, make sure we resolve each unparsed node
in the set before calling check relation on the outer LHS field.
Call it with strict == false because literals are allowed.

The latter changes the test added for #19538 a bit because
"22-33-" now gets a somewhat more useful error message about being
a possible but truncated Ethernet address instead of "not a field"
in a case where it could be a literal Ethernet address.

Add a test of the "in" operator where the RHS is a field.

Fix #20183