rtcp: Fix packet length
In RTP and RTCP, the packet length encoded in the packet is one less
than the number of 32 bit words of the length (with the exception of
certain Microsoft extensions), not the number of bytes. This was lost
in a conversion to a dissector table.

Fixup e5560357285ba8ee6b558fa82666fde2a6efc8e9

Fix #20188