Qt: Don't use theme stock icons on Windows or macOS
Qt 6.7 added QIcon::ThemeIcon with support for some, but not all,
of the FreeDesktop standard stock icons. This is ok in some cases,
theme icons next to our icons can look jarring, especially when, e.g.
the Windows theme provides "go-previous" and "go-next" but not
"go-jump", "go-first", and "go-last."

For now, disable theme icons on Windows and macOS and just use our
icons. This makes 6.7 and later look the same as 6.6 and earlier.