QUIC: Support initial_max_path 0 or 1 (draft-09 or later)
QUIC Multipath draft-07: "The value of the initial_max_paths parameter
MUST be at least 2.  An endpoint that receives a value less than 2
MUST close the connection with an error of type TRANSPORT_PARAMETER_ERROR."

QUIC Multipath draft-09 and later changed this to: "For example, if
initial_max_path_id is set to 1, only connection IDs associated with
Path IDs 0 and 1 should be issued by the peer. If an endpoint receives
an initial_max_path_id transport parameter with value 0, the peer aims
to enable the multipath extension without allowing extra paths immediately."

So consider multipath enabled by a peer regardless of the value for
initial_max_path_id draft-09 and draft-11, unlike in draft-07.