ftype-time: Fix multiplication
This still likely overflows on truly enormous values, but this is
correct for most common uses.

Do integer division with floating point for now to at least make
it largely accurate (as opposed to before.)

Fix one of the tests that gives a different (correct) result now
compared to before.

Fix #19150