DNP3: Only add reassembled AL packets in the FINAL fragment
It's possible to have multiple DNP3 AL fragments in the same
TCP segment. (or UDP datagram?)

Instead of checking the packet number and layer number (the last
of which is not useful, as tcp_dissect_pdus doesn't increase the
layer number for each pdu), simply check if the Final bit is set.
This prevents adding the reassembled packet multiple times in such
cases. (HTTP/2 has a similar approach.)

Also, if this isn't the first Data Link segment (PDU) then add a
separator when adding additional DNP3 transport information to
the Info column.

Fix #14436