dumpcap: Initialize failure message routines
Initialize a set of failure message routines for dumpcap, so that
if any code in libwsutil calls one of these routines, it prints a
message instead of segfaulting. Put these routines in libwsutil,
not libui, as dumpcap intentionally doesn't link against libui.

Some of the routines are capture file routines, where the parameters
are libwiretap constants that can only be properly interpreted using
libwiretap functions. dumpcap intentionally doesn't link against
libwiretap either, so these functions should never be called.

It might make sense to move those libwiretap-related report message
routines from libwsutil to libwiretap, though as those are exported
that would be an API break and couldn't be done in a backport.

Fix #20108