lemon: Suppress type-limits warning on gcc with generated code
After updating gcc, the lemon generated code starting giving
warnings with -Wtype-limits (enabled by -Wextra):

./ascend_parser.c: In function ‘AscendParserFinalize’:
./ascend_parser.c:866:21: warning: comparison is always true due to limited range of data type [-Wtype-limits]

The code in question:
    if( yytos->major>=YY_MIN_DSTRCTR ){

compares a type #defined as an unsigned char to a value #defined as 0.
Since there is no minimum symbol value to have a destructor, it's
always true, but that's by intent.

Ping #20175