lua: Load bit and rex_pcre2 as though through require
Use luaL_requiref to load the bit and rex_pcre2 modules, which
adds them to the package.loaded table as though through require,
as Lua does with its included standard libraries.

This causes `require("bit")` and `require("rex_pcre2")` statements
in Lua dissectors to behave as expected, if albeit usually unecessary.
The tables are still added to the global environment as before.

Fix #20213