lua: Add a fixture to test Lua scripts under different locales
Add a cross-platform way to run a Lua script under a different
locale by calling os.setlocale.

On UN*X we could have a different test_env fixture with LC_ environment
variables to override the native locale, but that doesn't seem to work
on Windows.

Add some tests to util.lua for setting range preferences with
comma separated ranges.

Change the way the dissector is added to the DissectorTable in
protofield.lua so that it uses a comma separated range. Note this
test will fail if only one of the ports is set (as happened silently
with Lua 5.2 and earlier in locales using decimal comma).

Run these tests with a German locale, so that the decimal and thousands
separators are swapped from English.

(cherry picked from 6e2e4440430b4b57e926d7838af92a7c9ff13193)