DHCP: Handle rfc3396 long options in a generic manner
Handle RFC 3396 Long Option concatenation in a generic manner and
according to the RFC.

Per RFC 3396 "[t]he aggregate option buffer is made up of the
optional parameters field, the file field, and the sname field,
in that order. WARNING: This is not the physical ordering of
these fields in the DHCP packet," and that options can be split
across the fields ("if an option were split into three parts and
each part went into one of the possible option fields, the first
part would go into the optional parameters field, the second part
would go into the file field, and the third part would go into the
sname field.")

This means that unlike the current handling for Options 119 and 120,
the indices should not be reset across those fields. It also means
that the overloaded fields cannot be processed inside the Option
Overload (Option 52), but must be processed after the entire optional
parameters field.

Further note that RFC 3396 4 Applicability and 7 Decoding Agent Behavior
both clearly indicate that decoding agents MUST consider any case where
an option code is repeated as a split option and reassemble it, even for
options that are not "concatenation-requiring."

However, cf. the very recent Internet-Draft
https://www.ietf.org/archive/id/draft-tojens-dhcp-option-concat-considerations-00.html
which notes that real world DHCP implementations do not do this, and
do not have RFC 3396 concenation handling for options that do not
support concatenation, especially fixed length options.

So we may wish to concatnate only certain options, or have a range
preference for which options are concatenated.

Tested on the sample captures, including the samples with option
overloading.

Fix #20201