GTP: Retrieve gtp_conv_info inside GTP' data transfer resp
args is only non-null when Session Tracking is on, which is not
by default.

Retrieve (creating if necessary) the gtp conversation info inside
the GTP' Data Transfer Response instead of passing it in args.

Make a note about how the source IP address of the response does not
have to be the destination IP of the address, though the other 3
ports and addresses do have to match. (See 3GPP 29.060 10.1 UDP/IP;
I think we do have some sample captures somewhere that show this.)

Fix #20222