Qt: Make I/O Graphs Average over Time backwards compatible
Move the new AOT field to the end of the UAT description, and add
it to iog_uat_defaults.

Use header()->moveSection() to change the default *visual* location
of the column instead of where it is stored in the UAT. This allows
I/O Graphs for earlier versions to still work on master. (It will
cause errors for any graphs on master between
commit 4a4351e8e21966e13d2cf11abb8c02ffaf37ce4e and now.)