Clean up About info.
Rename "Compiler info" to "Compile-time info", as it includes more
information than the compiler and the bit-width for which we're
compiling.

In particular, GLib is *not* a compiler library, so show its version the
same way the version is shown in the run-time info, as a separate item.

Indent the "with" and "without" information with a one-space indent for
both the "With:" header and the "Without:" header.