capture info: Store total counts as 64 bit integers
Wireshark can't handle more than 2^31-1 packets in the packet list
(because QTreeView uses an int to count the number of rows), and
tshark can't handle more than 2^32-1 packets.

However, dumpcap is not limited to capturing only a 32 bit number
of packets, and if the packet list is not being updated in real time,
and multiple files are used so that no file has too many packets to
open, it's possible to capture more than 2^32 packets in a single
session without issues. When capture stops, the last file will be
opened, but not until then.

The capture info dialog ought to still work in such a situation.

cap_session->count can be a uint64_t for the same reason (it
only should get that large in a "don't update the packet list in
real time" session) so that the main status bar also can handle big
numbers.

Fix #20233