lemon: Suppress MSVC warning via pragma, not in CMake
MSVC requires a very specific ordering of compiler warning flags,
and CMake doesn't comply with it, which means that if you change
the level of a warning and then try to disable it via CMake, it
doesn't actually get disabled.

https://gitlab.kitware.com/cmake/cmake/-/issues/18736

Disabling it via a pragma when compiling through ws_diag_control.h
does suppress the warning.

Follow up to 7993b6836c09f20e7c117672582399bb327a8711