wiretap: don't use void parameter type in wtap_new_ipv6_callback_t
The wtap_new_ipv6_callback_t was initially introduced to allow
add_new_ipv6 to be called from the wiretap library. It's first parameter
is a 128-bit IPv6 address, so let's use that instead of a void pointer.

This fixes an UBSan error that could cause a crash when CFI is enabled:

    wiretap/wtap.c:1623:4: runtime error: call to function add_ipv6_name through pointer to incorrect function type 'void (*)(const void *, const char *, bool)'

Link: https://maskray.me/blog/2022-12-18-control-flow-integrity#fsanitizefunction