VJC: Improve connection tracking across passes
Use a `wmem_multimap` to track information associated with a connection
ID rather than depending on a `conversation`. The use of
`wmem_multimap_lookup32_le()` gives more consistent results when a conn
ID can have a variety of compressed and uncompressed headers.

Use a `wmem_map` to track the conn ID of compressed packets lacking this
field.

Other minor improvements:

* Replace bespoke `GET_16()` etc with standardized `pntoh16()` etc.
* Switch `if` conditions to use consistent order.
* Remove redundant word "flag" from label of flag fields.
* Ran through `tools/check_dissector.py`.