endpoint_dialog: small preference-fetching code cleanup.
Fetch the ip.conv_agg_flag preference value with prefs_get_bool_value(),
after getting the pref_t for that preference with
prefs_find_preference().

That means we fetch a Boolean preference with the right routine, and
that we only fetch a pointer to the "ip" preferences module once (rather
than finding it and then, in a call to prefs_get_uint_value(), finding
it again - prefs_find_preference() checks whether it's handed a null
module pointer and, if so, returns null).