Qt: Get the debounce timer value from the prefs structure.
We do so in ui/qt/widgets/capture_filter_edit.cpp; we might as well do
so in the other files that use it.

That simplifies the process of fetching it, and lets us get rid of the
no-longer-used prefs_get_uint_value().  That lets us name the routine to
fetch unsigned integer preference values similarly to the routines to
fetch most other preference values, by removing the _real from the name.