Qt: Add a setType function to DisplayFilterEdit for use with ui files
Qt ui files only call a constructor with a QWidget first parameter
and can't pass other parameters or call other constructors.
Add a function to set the type after initial construction that can
be called in the constructors of other widgets after setupUi.

This makes it possible to use other DisplayFilterEditTypes in ui files.