dfilter: Do not crash on vals() with incompatible fields with shared abbrevs
Even though we check in the semantic check that an abbreviation can be
be converted, we have to check when actually converting because some
abbreviations are shared with multiple incompatible fields.

For the same reason, the semantic check perhaps should allow the filter
if at least one field registration works, not just the last. But really
we should enforce not having incompatible fields.