wslua: Throw an error properly for TreeItem:add with nil value
The documentation for treeitem:add has long said:

This function has a complicated form:... such that if the first argument is a
ProtoField or a Proto, the second argument is a TvbRange, and a third argument
is given, it’s a value; but if the second argument is a non-TvbRange, then
it’s the value (as opposed to filling that argument with 'nil', *which is
invalid for this function*). If the first argument is a non-ProtoField and a
non-Proto then this argument can be either a TvbRange or a label, and the value
is not in use.

Go ahead and check for the nil value and create a more user-friendly error
message than what Lua will give so it's easier to spot the error. Lua
will always error here.

The Lua error also creates some issues with certain forms of
postdissectors and protocols calling protocols because of Lua's
own setjmp/longjmp using exception handler in C interfering with
the Wireshark kazlib exception handling, causing jumping to the
wrong location and accessing something that might not be valid anymore.
See 8181ec3463631536f81b9aa568f887fa8dc2a10a

Fix #20244