cllog: Add a file-format decoding test.
Added a trivial CL2000 capture file format sample, as obtained from:

    https://canlogger.csselectronics.com/clx000-docs/cl2000/log/index.html

The cllog parser is modified to recognize the "Logger type" name in
that sample capture, which differs slightly from that already known to
the parser ("CL2000" versus "CANLogger2000").  Corresponding changes
are made for the CL1000 capture file, also available on the above
website -- but not added here because it provides no additional
coverage beyond the "Logger type" value, and for CL3000 just for the
sake of consistency.

Add a python testcase to read the added file.

Additional testing with two further sample captures available from
https://canlogger.csselectronics.com/clx000-intro/log-file-tools/
revealed a defect which caused file parsing to terminate as soon as
any packet was shorter than 8 bytes; Fix it by accepting NUL to
terminate the packet (in addition to CR/LF).